+++ MelonLoader/InternalUtils/UnityInformationHandler.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Runtime.CompilerServices;
@@ -11,13 +13,14 @@ namespace MelonLoader.InternalUtils
     {
         public static string GameName { get; private set; } = "UNKNOWN";
         public static string GameDeveloper { get; private set; } = "UNKNOWN";
-        public static UnityVersion EngineVersion { get; private set; }
+        public static UnityVersion EngineVersion { get; private set; } = new UnityVersion();
         public static string GameVersion { get; private set; } = "0";
 
         internal static void Setup()
         {
             string gameDataPath = MelonUtils.GetGameDataDirectory();
             AssetsManager assetsManager = new AssetsManager();
+            assetsManager.LoadIncludedClassPackage();
 
             ReadGameManager(gameDataPath, assetsManager);
             SetDefaultConsoleTitleWithGameName(GameName, GameVersion);
@@ -32,32 +35,49 @@ namespace MelonLoader.InternalUtils
 
         private static void ReadGameManager(string gameDataPath, AssetsManager assetsManager)
         {
-            string bundlePath = Path.Combine(gameDataPath, "globalgamemanagers");
-            if (!File.Exists(bundlePath))
-                bundlePath = Path.Combine(gameDataPath, "mainData");
+            AssetsFileInstance instance = null;
+            try
+            {
+                string bundlePath = Path.Combine(gameDataPath, "globalgamemanagers");
+                if (!File.Exists(bundlePath))
+                    bundlePath = Path.Combine(gameDataPath, "mainData");
 
-            if (!File.Exists(bundlePath))
-                return;
+                if (!File.Exists(bundlePath))
+                    return;
 
-            AssetsFileInstance instance = assetsManager.LoadAssetsFile(bundlePath, true);
-            if (instance == null)
-                return;
+                instance = assetsManager.LoadAssetsFile(bundlePath, true);
+                if (instance == null)
+                    return;
 
-            assetsManager.LoadIncludedClassPackage();
-            assetsManager.LoadClassDatabaseFromPackage(instance.file.typeTree.unityVersion);
+                assetsManager.LoadClassDatabaseFromPackage(instance.file.typeTree.unityVersion);
+                EngineVersion = UnityVersion.Parse(instance.file.typeTree.unityVersion);
 
-            EngineVersion = UnityVersion.Parse(instance.file.typeTree.unityVersion);
+                List<AssetFileInfoEx> assetFiles = instance.table.GetAssetsOfType(129);
+                if (assetFiles.Count > 0)
+                {
+                    AssetFileInfoEx playerSettings = instance.table.GetAssetsOfType(129).First();
+                    AssetTypeInstance assetTypeInstance = assetsManager.GetTypeInstance(instance, playerSettings);
+                    if (assetTypeInstance != null)
+                    {
+                        AssetTypeValueField playerSettings_baseField = assetTypeInstance.GetBaseField();
 
-            AssetFileInfoEx buildSettings = instance.table.GetAssetsOfType(129).First();
-            AssetTypeValueField buildSettings_baseField = assetsManager.GetTypeInstance(instance, buildSettings).GetBaseField();
-            GameVersion = buildSettings_baseField.Get("bundleVersion").GetValue().AsString();
+                        AssetTypeValueField bundleVersion = playerSettings_baseField.Get("bundleVersion");
+                        if (bundleVersion != null)
+                            GameVersion = bundleVersion.GetValue().AsString();
 
-            AssetFileInfoEx playerSettings = instance.table.GetAssetsOfType(129).First();
-            AssetTypeValueField playerSettings_baseField = assetsManager.GetTypeInstance(instance, playerSettings).GetBaseField();
-            GameDeveloper = playerSettings_baseField.Get("companyName").GetValue().AsString();
-            GameName = playerSettings_baseField.Get("productName").GetValue().AsString();
+                        AssetTypeValueField companyName = playerSettings_baseField.Get("companyName");
+                        if (companyName != null)
+                            GameDeveloper = companyName.GetValue().AsString();
 
-            instance.file.Close();
+                        AssetTypeValueField productName = playerSettings_baseField.Get("productName");
+                        if (productName != null)
+                            GameName = productName.GetValue().AsString();
+                    }
+                }
+            }
+            catch(Exception ex) { MelonLogger.Error(ex); }
+            if (instance != null)
+                instance.file.Close();
         }
 
         [MethodImpl(MethodImplOptions.InternalCall)]